--- domain/src/main/java/com/nextstep/edu/tictactoe/domain/TictactocPut.kt
@@ -1,9 +1,0 @@
-package com.nextstep.edu.tictactoe.domain
-
-import com.nextstep.edu.tictactoe.domain.model.GameResult
-import com.nextstep.edu.tictactoe.domain.model.Point
-
-interface TictactocPut {
-
-    fun put(point: Point): GameResult
-}
