+++ domain/src/main/java/com/nextstep/edu/tictactoe/domain/model/Behavior.kt
@@ -1,0 +1,10 @@
+package com.nextstep.edu.tictactoe.domain.model
+
+data class RandomBehavior(
+    val behavior: Behavior,
+    val point: Point
+)
+
+enum class Behavior {
+    INTERRUPT, WIN, UNKNOWN
+}
